update tool ProcessFileList with open features
before all features were propagated. since some features can be dead (not in the list of open features) this lead to false positives. now we can explicitly provide a list of open features and all features not in this list do not make it into the presence condition.

additionally introduced the optional library here, see
https://github.com/jwebb/optional